Encapsulate ActiveCalculator into Package `.calc`
Changes
- Move `Expression` `Operand` `Operator` models to `.calc.model`
- Remove dependencies from `.calc.*` to `PanelInput`
- ActiveCalculatorViewModel relays `expression` and `evaluationResult`

Todo
- Remove dependency from ActiveCalculator to PanelInput
