+++ app/src/main/java/io/pjhjohn/calculator/model/PanelInput.kt
@@ -1,5 +1,7 @@
 package io.pjhjohn.calculator.model
 
+import io.pjhjohn.calculator.calc.model.Operator
+
 sealed class PanelInput(val type: Type, val value: String) {
 
     object Number0 : PanelInput(Type.NUMBER, "0")
@@ -25,9 +27,31 @@ sealed class PanelInput(val type: Type, val value: String) {
         NUMBER, OPERATOR, RESET, EVALUATE
     }
 
+    fun toOperator(): Operator = when (this) {
+        OperatorPlus -> Operator.Plus
+        OperatorMinus -> Operator.Minus
+        OperatorMultiply -> Operator.Multiply
+        OperatorDivide -> Operator.Divide
+        else -> throw IllegalArgumentException("Cannot convert PanelInput($this) to Operator")
+    }
+
+    fun toInt(): Int = when (this) {
+        Number0 -> 0
+        Number1 -> 1
+        Number2 -> 2
+        Number3 -> 3
+        Number4 -> 4
+        Number5 -> 5
+        Number6 -> 6
+        Number7 -> 7
+        Number8 -> 8
+        Number9 -> 9
+        else -> throw IllegalArgumentException("Cannot convert PanelInput($this) to Int")
+    }
+
 }
 
-fun String.asPanelInput(): PanelInput = when (this) {
+fun String.toPanelInput(): PanelInput = when (this) {
     PanelInput.Number0.value -> PanelInput.Number0
     PanelInput.Number1.value -> PanelInput.Number1
     PanelInput.Number2.value -> PanelInput.Number2
